feat: add punch behaviour to priest
Punch animation is working when input Fire1 is triggered. The priest punches the last direction he was moving so it's possible to punch in 4 directions left/right/up/down.
There is a circle collider following punch frame animation so it's possible to test when enemies are hit by a punching priest.
It's lacking the punch hit test and the damage calculations.